+++ packages/map-template/src/components/MapTemplate/MapTemplate.jsx
@@ -235,8 +235,8 @@ function MapTemplate({ apiKey, gmApiKey, mapboxAccessToken, venue, locationId, p
 
             const miSdkApiTag = document.createElement('script');
             miSdkApiTag.setAttribute('type', 'text/javascript');
-            miSdkApiTag.setAttribute('src', 'https://app.mapsindoors.com/mapsindoors/js/sdk/4.46.0/mapsindoors-4.46.0.js.gz');
-            miSdkApiTag.setAttribute('integrity', 'sha384-Ygw6sO8Q71z/FxztdpmuOo+hYFf5ozKSIrrm+MDsIaRapLjN7STzmbVNEfHegsqt');
+            miSdkApiTag.setAttribute('src', 'https://app.mapsindoors.com/mapsindoors/js/sdk/4.47.1/mapsindoors-4.47.1.js.gz');
+            miSdkApiTag.setAttribute('integrity', 'sha384-O4P7Q8ewNTpyVdRoIx/AbR/eMQBzwNHs5SU36XpC3LyrHSgijp/nVlliPINFGLxt');
             miSdkApiTag.setAttribute('crossorigin', 'anonymous');
             document.body.appendChild(miSdkApiTag);
             miSdkApiTag.onload = () => {