+++ packages/components/CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [13.32.1] - 2026-01-06
+
+### Fixed
+
+- Fixed `TypeError` in `mi-my-position` component when accessing compass button style before the element is rendered.
+
+## [13.32.0] - 2025-12-11
+
+### Added
+
+- `showFloor` flag for `<mi-location-info>` to conditionally display floor information.
+
 ## [13.27.1] - 2025-11-17
 
 ### Added